new changes
result.html  ( page ready) and transcript( just made this page no content has been added yet).

I have created a static 8 semester page but if we can do a dynamic ( only the done semesters are seen) i will try it today.

Tried to add the list ( went through 20 videos on the topic). I have applied a way of doing it which is working for all the videos but apparently not in this. Please check what needs to be done for that since i am lost now.